+++ src/components/cartitem/cartitem.css
@@ -3,7 +3,6 @@
   flex-direction: row;
   align-items: center;
   justify-content: space-between;
-  gap: 10px;
 
   width: 100%;
   height: 50px;
@@ -19,6 +18,12 @@
   text-transform: capitalize;
   margin: 0 10px;
   flex-grow: 2;
+  display: block;
+  width: 400px;
+}
+.item-container--righ-col {
+  display: flex;
+  flex-direction: row;
 }
 .qty-container {
   display: flex;
@@ -41,9 +46,11 @@
   text-align: center;
 }
 .price {
+  text-align: center;
   font-weight: 600;
   font-size: 22px;
   margin: 0 10px;
+  width: 4em;
 }
 .delete-item {
   margin-right: 10px;
