+++ src/components/cartpage/cartpage.css
@@ -1,0 +1,12 @@
+.cart-page--hero {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  height: calc(100vh - 70px);
+}
+.cart-page--cart-items-container {
+  font-weight: 600;
+  font-size: 22px;
+  margin-top: 50px;
+}
